+++ controllers/leg_controller.go
@@ -196,6 +196,14 @@ func GetStatisticsForLeg(w http.ResponseWriter, r *http.Request) {
 			return
 		}
 		json.NewEncoder(w).Encode(stats)
+	} else if match.MatchType.ID == models.JDCPRACTICE {
+		stats, err := data.GetJDCPracticeStatisticsForLeg(legID)
+		if err != nil {
+			log.Println("Unable to get JDC Practice statistics", err)
+			http.Error(w, err.Error(), http.StatusInternalServerError)
+			return
+		}
+		json.NewEncoder(w).Encode(stats)
 	} else {
 		stats, err := data.GetX01StatisticsForLeg(legID)
 		if err != nil {