+++ controllers/match_controller.go
@@ -290,6 +290,14 @@ func GetStatisticsForMatch(w http.ResponseWriter, r *http.Request) {
 			return
 		}
 		json.NewEncoder(w).Encode(stats)
+	} else if match.MatchType.ID == models.JDCPRACTICE {
+		stats, err := data.GetJDCPracticeStatisticsForMatch(matchID)
+		if err != nil {
+			log.Printf("Unable to get JDC Practice statistics for match %d: %s", matchID, err)
+			http.Error(w, err.Error(), http.StatusInternalServerError)
+			return
+		}
+		json.NewEncoder(w).Encode(stats)
 	} else {
 		stats, err := data.GetX01StatisticsForMatch(matchID)
 		if err != nil {