+++ controllers/player_controller.go
@@ -282,6 +282,15 @@ func GetPlayerMatchTypeStatistics(w http.ResponseWriter, r *http.Request) {
 		json.NewEncoder(w).Encode(stats)
 		return
 
+	case models.JDCPRACTICE:
+		stats, err := data.GetJDCPracticeStatisticsForPlayer(id)
+		if err != nil {
+			log.Println("Unable to get JDC Practice Statistics for player", err)
+			http.Error(w, err.Error(), http.StatusInternalServerError)
+			return
+		}
+		json.NewEncoder(w).Encode(stats)
+		return
 	default:
 		log.Println("Unknown match type parameter")
 		http.Error(w, err.Error(), http.StatusBadRequest)
@@ -427,6 +436,15 @@ func GetPlayerMatchTypeHistory(w http.ResponseWriter, r *http.Request) {
 		legs, err := data.GetGotchaHistoryForPlayer(id, limit)
 		if err != nil {
 			log.Println("Unable to get Gotcha history for player", err)
+			http.Error(w, err.Error(), http.StatusInternalServerError)
+			return
+		}
+		json.NewEncoder(w).Encode(legs)
+		return
+	case models.JDCPRACTICE:
+		legs, err := data.GetJDCPracticeHistoryForPlayer(id, limit)
+		if err != nil {
+			log.Println("Unable to get JDC Practice history for player", err)
 			http.Error(w, err.Error(), http.StatusInternalServerError)
 			return
 		}