+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [unreleased] - TBD
+#### Feature
+- Support for new game type `JDC Practice Routine`
+
 ## [2.0.0] - 2021-09-19
 #### Feature
 - Support for new game type `Kill Bull`
@@ -46,6 +50,7 @@
 #### Feature
 - Intial version of API for [kcapp-frontend](https://github.com/kcapp/frontend)
 
+[unreleased]: https://github.com/kcapp/api/compare/v2.0.0...develop
 [2.0.0]: https://github.com/kcapp/api/compare/v1.2.0...v2.0.0
 [1.2.0]: https://github.com/kcapp/api/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/kcapp/api/compare/v1.0.0...v1.1.0